AVOptions -- modified and deleted files * keep only functions that "work" * remove unrelated and supplementary material
