+++ test/runtests.jl
@@ -1,2 +1,9 @@
-#include("avio.jl")
-include("avoptions_tests.jl")
+include("avio.jl")
+
+@osx_only begin
+    include("avoptions_tests.jl")
+end
+
+@windows_only begin
+    include("avoptions_tests.jl")
+end
